+++ src/main/java/com/bv/pet/jeduler/repositories/StatisticsRepository.java
@@ -3,5 +3,5 @@ package com.bv.pet.jeduler.repositories;
 import com.bv.pet.jeduler.entities.Statistics;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatisticsRepository extends JpaRepository<Statistics, Long> {
+public interface StatisticsRepository extends JpaRepository<Statistics, Short> {
 }
