+++ src/main/java/com/bv/pet/jeduler/services/category/ICategoryService.java
@@ -8,5 +8,5 @@ public interface ICategoryService {
     List<CategoryDto> all();
     CategoryDto create(CategoryDto categoryDto);
     CategoryDto update(CategoryDto categoryDto);
-    void delete(Long id);
+    void delete(Short id);
 }
